More on parallel indexing.  The code doesn't yet work.
git-svn-id: https://groke.otago.ac.nz/ant/ant@757 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/memory_indexer_node.h
@@ -16,7 +16,7 @@ class ANT_memory_indexer_node
 {
 public:
 	ANT_string_pair string;
-	long term_frequency;
+	long long term_frequency;
 
 public:
 	void *operator new(size_t count, ANT_memory *memory) { return memory->malloc(count); }
